#ZEX#fixes#add & get refresh token

+++ src/DB.Api/Application/Validators/Commands/CreateUserCommandValidator.cs
@@ -36,7 +36,7 @@ namespace DB.Api.Application.Validators.Commands
                 .EmailAddress(EmailValidationMode.AspNetCoreCompatible)
                 .WithMessage("Do not match format");
             RuleFor(command => command.Type)
-                .GreaterThanOrEqualTo(1)
+                .GreaterThanOrEqualTo((short)1)
                 .When(w => w.Type.HasValue)
                 .WithMessage("Must be greater than or equal to 1");
         }
